Update getRule to observe the time zones DST switching times are given in
Zone file rules contain a parameter 'AT' which specifies the time of day
the rule starts to be active.
This parameter can be given in UTC, or local wallclock time, or local
standard time.
By now, only wallclock time was handled correctly. With this patch, all
three possibilities are supported.